requirements.txt
Django==2.2
django-debug-toolbar==2.2
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-extensions==0.6.0
future==0.18.2
PyJWT==1.7.1
pytz==2019.3
serializers==0.2.4
sqlparse==0.3.0
